Include disk usage (Data.fs and blobstorage) in stats:
Disk usage stats are gathered using the `du` utility. In order
to avoid running an I/O heavy and potentially long running process
like `du` from inside a Zope instance, this is being done from the
bin/dump-content-stats wrapper script.

This script invokes `du`, stores the results in a JSON file, and
then triggers the @@dump-content-stats view as usual.

That view then picks up the disk usage stats from that file
(if present), and includes them in the dumped stats.